requirements.txt
requests
pandas
numpy
scipy
matplotlib
scikit-learn
pyarrow
pyyaml
markdown